+++ packages/eslint-plugin-prairielearn/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/eslint-plugin
 
+## 2.1.6
+
+### Patch Changes
+
+- d97b97a: Upgrade all JavaScript dependencies
+
 ## 2.1.5
 
 ### Patch Changes