+++ packages/flash/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/flash
 
+## 2.0.18
+
+### Patch Changes
+
+- d97b97a: Upgrade all JavaScript dependencies
+
 ## 2.0.17
 
 ### Patch Changes